rename and revise create script
